infra/cdk/requirements.txt
aws-cdk-lib>=2.110.0
constructs>=10.3.0
aws-cdk.aws-eks>=2.110.0
aws-cdk.aws-ec2>=2.110.0
aws-cdk.aws-iam>=2.110.0
aws-cdk.aws-dynamodb>=2.110.0
aws-cdk.aws-s3>=2.110.0
aws-cdk.aws-bedrock>=2.110.0
pyyaml>=6.0